requirements.txt
# Streamlit UI
streamlit>=1.28.0

# LLM providers
openai>=1.0.0
anthropic>=0.40.0

# Image generation
replicate>=0.25.0
requests>=2.28.0

# Text-to-speech
kokoro>=0.9.4
soundfile>=0.12.0

# Video assembly
moviepy>=2.0.0
Pillow>=10.0.0
numpy>=1.24.0
pycairo>=1.26.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Testing
pytest>=8.0.0
pytest-cov>=5.0.0
pytest-timeout>=2.3.0
pytest-xdist>=3.6.0
scikit-image>=0.24.0
